Workaround WFLY-13345
Extract the `AsyncListener` from `TimedHandler` to work around
https://issues.redhat.com/browse/WFLY-13345
